pyproject.toml
[project]
name = "bingo-card-generator"
version = "1.0.0"
description = "A customizable bingo card generator that creates interactive HTML bingo cards"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Bingo App Developer" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Games/Entertainment",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "jinja2",
    "pillow",
    "click",
    "rich",
    "questionary",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "black>=23.0",
]

[project.scripts]
create-bingo-card = "create_bingo_card:main"
create-spooky-bingo = "create_bingo_card:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "*.jinja",
    "*.csv",
    "images/*",
]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

# Mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, can be strict later
ignore_missing_imports = true

# Black configuration
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
